SpartaMonitor/requirements.txt
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.59.0
GPUtil==1.4.0
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==10.4.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
ttkbootstrap==1.14.2

# 🔥 Added for temperature monitoring (Windows only)
wmi==1.5.1; platform_system=="Windows"
pywin32==308; platform_system=="Windows"
